pdf-compressor/requirements.txt
Flask==2.2.2
google-cloud-storage==2.7.0
google-cloud-firestore==2.9.1
requests==2.28.2
gunicorn==20.1.0
